backend/requirements.txt
flask
flask-cors
pillow
numpy
pytest
opencv-python
requests
imageio